Move ResolutionControl
