+++ front/app/components/admin/ResolutionControl/index.tsx
@@ -1,9 +1,10 @@
 import React, { PureComponent } from 'react';
 import styled from 'styled-components';
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from '../messages';
-import { IResolution } from '..';
+import messages from './messages';
 import Tabs from 'components/UI/Tabs';
+
+export type IResolution = 'day' | 'week' | 'month';
 
 const Container = styled.div``;
