+++ front/app/components/admin/ResolutionControl/messages.ts
@@ -1,0 +1,16 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  resolutionday: {
+    id: 'app.containers.AdminPage.DashboardPage.resolutionday',
+    defaultMessage: 'in Days',
+  },
+  resolutionweek: {
+    id: 'app.containers.AdminPage.DashboardPage.resolutionweek',
+    defaultMessage: 'in Weeks',
+  },
+  resolutionmonth: {
+    id: 'app.containers.AdminPage.DashboardPage.resolutionmonth',
+    defaultMessage: 'in Months',
+  },
+});
