+++ front/app/containers/Admin/dashboard/index.tsx
@@ -206,9 +206,6 @@ export const GraphCardFigureChange = styled.span`
 `;
 
 export type IGraphUnit = 'users' | 'ideas' | 'comments' | 'votes';
-
-export type IResolution = 'day' | 'week' | 'month';
-
 interface Props {
   authUser: GetAuthUserChildProps;
   children: JSX.Element;
