+++ front/app/modules/commercial/insights/admin/containers/Reports/charts/LineBarChart.tsx
@@ -34,7 +34,6 @@ import {
 } from 'recharts';
 import {
   IGraphUnit,
-  IResolution,
   GraphCard,
   NoDataContainer,
   GraphCardInner,
@@ -44,6 +43,7 @@ import {
   GraphCardFigure,
   GraphCardFigureChange,
 } from './';
+import { IResolution } from 'components/admin/ResolutionControl';
 import { Popup } from 'semantic-ui-react';
 import { Icon } from 'cl2-component-library';
